Removed AuthItem
